Rename reports.component to reports-layout.component, move reports into pages folder
