+++ apps/web/src/app/reports/pages/weak-passwords-report.component.html
